+++ src/eval_framework/tasks/benchmarks/balancedcopa.py
@@ -1,11 +1,12 @@
 from eval_framework.tasks.benchmarks.copa import COPA
 
 
-class BalancedCOPA(COPA[str]):
-    """COPA dataset: https://huggingface.co/datasets/pkavumba/balanced-copa"""
+class BalancedCOPA(COPA):
+    """Balanced-COPA dataset: https://huggingface.co/datasets/pkavumba/balanced-copa"""
 
     NAME = "BalancedCOPA"
-    DATASET_PATH = "aps/balanced_copa"
+    DATASET_PATH = "pkavumba/balanced-copa"
     # The dataset combines train and validation splits, so we use test split for evaluation.
     SAMPLE_SPLIT = "test"  # 500 examples
     FEWSHOT_SPLIT = "test"  # 500 examples
+    SUBJECTS = ["no_subject"]